requirements.deb11.txt
altgraph==0.17.2
anyio==3.6.1
astroid==2.12.4
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
fastapi==0.81.0
future==0.18.2
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
pefile==2022.5.30
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.9.2
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.9
pylint==2.15.0
python-vagrant==0.5.15
pywin32-ctypes==0.2.0
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
wrapt==1.14.1
